# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API for Gopher Social, a social media platform for Gophers.",
        "title": "Gopher Social API",
        "contact": {
            "name": "Rohit Vilas Ingole",
            "email": "rohit.vilas.ingole@gmail.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://github.com/DataRohit/Gopher-Social-Backend/blob/master/license"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/post/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new post by a logged-in user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "Create a new post",
                "parameters": [
                    {
                        "description": "Request Body for creating a post",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreatePostPayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created post",
                        "schema": {
                            "$ref": "#/definitions/models.CreatePostSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid input",
                        "schema": {
                            "$ref": "#/definitions/models.CreatePostErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - User not logged in or invalid token",
                        "schema": {
                            "$ref": "#/definitions/models.CreatePostErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - User account is inactive or banned",
                        "schema": {
                            "$ref": "#/definitions/models.CreatePostErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Failed to create post",
                        "schema": {
                            "$ref": "#/definitions/models.CreatePostErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/post/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a list of posts created by the logged-in user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "List posts of logged-in user",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number for pagination",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved list of user's posts",
                        "schema": {
                            "$ref": "#/definitions/models.ListMyPostsSuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - User not logged in or invalid token",
                        "schema": {
                            "$ref": "#/definitions/models.ListMyPostsErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Failed to fetch user's posts",
                        "schema": {
                            "$ref": "#/definitions/models.ListMyPostsErrorResponse"
                        }
                    }
                }
            }
        },
        "/post/user/{identifier}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a list of posts created by a user identified by username, email, or user ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "List posts by user identifier",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User Identifier (username, email, or user ID)",
                        "name": "identifier",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number for pagination",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved list of user's posts",
                        "schema": {
                            "$ref": "#/definitions/models.ListUserPostsSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid input",
                        "schema": {
                            "$ref": "#/definitions/models.ListUserPostsErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - User not logged in or invalid token",
                        "schema": {
                            "$ref": "#/definitions/models.ListUserPostsErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found - User not found",
                        "schema": {
                            "$ref": "#/definitions/models.ListUserPostsErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Failed to fetch user's posts",
                        "schema": {
                            "$ref": "#/definitions/models.ListUserPostsErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CreatePostErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.CreatePostPayload": {
            "type": "object",
            "required": [
                "content",
                "title"
            ],
            "properties": {
                "content": {
                    "type": "string",
                    "example": "This is the main content of my post."
                },
                "description": {
                    "type": "string",
                    "example": "A brief description of the post."
                },
                "sub_title": {
                    "type": "string",
                    "example": "A Catchy Subtitle"
                },
                "title": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 3,
                    "example": "My Awesome Post"
                }
            }
        },
        "models.CreatePostSuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Post Created Successfully"
                },
                "post": {
                    "$ref": "#/definitions/models.Post"
                }
            }
        },
...
        "models.ListMyPostsErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.ListMyPostsSuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "User Posts Retrieved Successfully"
                },
                "posts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Post"
                    }
                }
            }
        },
        "models.ListUserPostsErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.ListUserPostsSuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "User Posts Retrieved Successfully"
                },
                "posts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Post"
                    }
                }
            }
        },
        "models.Post": {
            "type": "object",
            "properties": {
                "author": {
                    "description": "Removed from response as per requirement",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.User"
                        }
                    ]
                },
                "content": {
                    "type": "string",
                    "example": "This is the main content of my post."
                },
                "created_at": {
                    "type": "string",
                    "example": "2025-01-25T12:34:01.159498Z"
                },
                "description": {
                    "type": "string",
                    "example": "A brief description of the post."
                },
                "dislikes": {
                    "type": "integer",
                    "example": 10
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "likes": {
                    "type": "integer",
                    "example": 100
                },
                "sub_title": {
                    "type": "string",
                    "example": "A Catchy Subtitle"
                },
                "title": {
                    "type": "string",
                    "example": "My Awesome Post"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2025-01-25T12:34:01.159498Z"
                }
            }
        },
...
        "models.Role": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Normal User"
                },
                "level": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "properties": {
                "banned": {
                    "type": "boolean",
                    "example": false
                },
                "created_at": {
                    "type": "string",
                    "example": "2025-01-25T12:34:01.159498Z"
                },
                "email": {
                    "type": "string",
                    "example": "john.doe@example.com"
                },
                "followers": {
                    "type": "integer"
                },
                "following": {
                    "type": "integer"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "is_active": {
                    "type": "boolean",
                    "example": false
                },
                "role": {
                    "$ref": "#/definitions/models.Role"
                },
                "timeout_until": {
                    "type": "string",
                    "example": "2025-01-25T12:34:01.159498Z"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2025-01-25T12:34:01.159498Z"
                },
                "username": {
                    "type": "string",
                    "example": "john_doe"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}